Created the completeGame() method in GameActivity to handle transition to score screen and handle adding the new word to the stored game data. This behaviour did not belong in the timer class.
[bugfix] Fixed bug where timer keeps running and invokes the score screen even if the user has left the game activity.
GameActivity has two new boolean flags; to track if the time has expired and to track if the GameActivity is visible on screen.
When the time expires, the Timer object now checks if the GameActivity is visible.
If yes, proceed to score screen; If no, set boolean timeUp flag to true. If user returns to the screen when the timeUp flag is true, the onResume() will instantiate the score screen class.